pyproject.toml
[project]
name = "hdx-python-geo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb",
    "geopandas",
    "httpx[http2]",
    "pyarrow",
    "python-dotenv",
    "tenacity",
]

[dependency-groups]
dev = ["ruff", "taskipy"]

[tool.taskipy.tasks]
esri = "python -m src.hdx.geo.esri"
parquet = "python -m src.hdx.geo.parquet"
export = "uv sync -q && uv export -q -o requirements.txt --no-dev --no-hashes"
ruff = "ruff format && ruff check && ruff format"
